Restructured Driver and DesiredCapability factory for Mobile & Web Test Session
 - Created WebDriverFactory and MobileDriverFactory for Web and Mobile
 - Removed DriverFactory and DesiredCapabilityFactory
 - Removed deprecated IPad and IPhone browser flavors and corresponding classes
 - Removed BrowserFlavors.GENERIC enum
