+++ client/src/main/java/com/paypal/selion/platform/grid/browsercapabilities/PhantomJSCapabilitiesBuilder.java
@@ -1,5 +1,5 @@
 /*-------------------------------------------------------------------------------------------------------------------*\
-|  Copyright (C) 2014 PayPal                                                                                          |
+|  Copyright (C) 2014-15 PayPal                                                                                       |
 |                                                                                                                     |
 |  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance     |
 |  with the License.                                                                                                  |
@@ -22,7 +22,7 @@ import org.openqa.selenium.remote.DesiredCapabilities;
 import com.paypal.selion.configuration.Config;
 import com.paypal.selion.configuration.Config.ConfigProperty;
 
-public class PhantomJSCapabilitiesBuilder extends DefaultCapabilitiesBuilder {
+class PhantomJSCapabilitiesBuilder extends DefaultCapabilitiesBuilder {
 
     @Override
     public DesiredCapabilities getCapabilities(DesiredCapabilities capabilities) {
